--- engine-core/src/main/kotlin/ru/cherryngine/engine/core/LoggerProvider.kt
@@ -1,15 +1,0 @@
-package ru.cherryngine.engine.core
-
-import io.micronaut.context.annotation.Factory
-import io.micronaut.context.annotation.Prototype
-import io.micronaut.inject.InjectionPoint
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-
-@Factory
-class LoggerProvider {
-    @Prototype
-    fun logger(injectionPoint: InjectionPoint<*>): Logger {
-        return LoggerFactory.getLogger(injectionPoint.declaringBean.declaringType.get())
-    }
-}
