+++ engine-core/src/main/kotlin/ru/cherryngine/engine/core/ext/DynamicRegistry.kt
@@ -1,0 +1,7 @@
+@file:Suppress("unused")
+
+package ru.cherryngine.engine.core.ext
+
+import net.minestom.server.registry.DynamicRegistry
+
+fun <T : Any> DynamicRegistry<T>.getId(value: T): Int? = getKey(value)?.let { getId(it) }
