+++ engine-core/src/main/kotlin/ru/cherryngine/engine/core/ext/Pos.kt
@@ -1,0 +1,9 @@
+@file:Suppress("unused")
+
+package ru.cherryngine.engine.core.ext
+
+import net.minestom.server.coordinate.Pos
+import ru.cherryngine.lib.math.View
+
+fun Pos.asView() = View(yaw, pitch)
+fun Pos.asQRot() = asView().getRotation()
