--- engine-core/src/main/kotlin/ru/cherryngine/engine/core/world/ChunkSupplier.kt
@@ -1,5 +1,0 @@
-package ru.cherryngine.engine.core.world
-
-fun interface ChunkSupplier {
-    fun create(blockHolder: BlockHolder, chunkX: Int, chunkZ: Int): Chunk?
-}
