+++ engine-core/src/main/kotlin/ru/cherryngine/engine/core/world/api/ChunkSupplier.kt
@@ -1,0 +1,7 @@
+package ru.cherryngine.engine.core.world.api
+
+import ru.cherryngine.engine.core.world.BlockHolder
+
+fun interface ChunkSupplier {
+    fun create(blockHolder: BlockHolder, chunkX: Int, chunkZ: Int): Chunk?
+}
