+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/models/Models.kt
@@ -1,7 +1,7 @@
 package ru.cherryngine.engine.scenes.models
 
 import ru.cherryngine.engine.core.server.ClientConnection
-import ru.cherryngine.engine.core.world.entity.EngineEntity
+import ru.cherryngine.engine.core.entity.EngineEntity
 import ru.cherryngine.engine.scenes.models.api.BodyPart
 
 sealed class Models {
