+++ engine-scenes/src/main/kotlin/ru/cherryngine/engine/scenes/models/builder/SkeletonModelBuilder.kt
@@ -2,7 +2,7 @@
 
 package ru.cherryngine.engine.scenes.models.builder
 
-import ru.cherryngine.engine.core.world.entity.EngineEntity
+import ru.cherryngine.engine.core.entity.EngineEntity
 import ru.cherryngine.engine.scenes.models.Models
 import ru.cherryngine.engine.scenes.models.api.BodyPart
